Show graph on canvas
This update contains a new function called "showGraphFromNeo4j" which
queries the database and displays the contents on the canvas.
Other functionality contained in this update is: new function
"getNodeById", changes to remove spaces from property names and
"neo4jNodeId" property is updated in database when a node is created.